+++ src/components/events/view-events/CalendarRow.test.js
@@ -1,30 +1,30 @@
 import React from "react";
-import CalendarView from "./CalendarView.js";
+import CalendarRow from "./CalendarRow.js";
 import { render } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
 import { createStore, applyMiddleware } from "redux";
 import { Provider } from "react-redux";
-import { rootReducer } from "../../utilities/reducers";
+import { rootReducer } from "../../../utilities/reducers";
 import thunk from "redux-thunk";
 
 import "@testing-library/jest-dom/extend-expect";
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
 
-describe("Test CalendarView static properties", () => {
-  let CalendarViewComponent;
+describe("Test CalendarRow static properties", () => {
+  let CalendarRowComponent;
   beforeEach(() => {
-    CalendarViewComponent = render(
+    CalendarRowComponent = render(
       <Provider store={store}>
         <BrowserRouter>
-          <CalendarView />
+          <CalendarRow />
         </BrowserRouter>
       </Provider>
     );
   });
 
-  test("CalendarView component renders", () => {
-    const firstDiv = document.querySelector(".calendar-view-main");
+  test("CalendarRow component renders", () => {
+    const firstDiv = document.querySelector(".calendar-row");
     expect(firstDiv.toBeInDocument);
   });
 });
