+++ src/components/events/view-events/CalendarView.js
