+++ src/components/events/view-events/EventDetails.js
