+++ src/components/events/view-events/StatusButton.test.js
@@ -1,30 +1,34 @@
 import React from "react";
-import EventDetails from "./EventDetails.js";
+import StatusButton from "./StatusButton.js";
 import { render } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
 import { createStore, applyMiddleware } from "redux";
 import { Provider } from "react-redux";
-import { rootReducer } from "../../utilities/reducers";
+import { rootReducer } from "../../../utilities/reducers";
 import thunk from "redux-thunk";
 
 import "@testing-library/jest-dom/extend-expect";
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
 
-describe("Test EventDetails static properties", () => {
-  let EventDetailsComponent;
+describe("Test StatusButton static properties", () => {
+  let StatusButtonComponent;
   beforeEach(() => {
-    EventDetailsComponent = render(
+    StatusButtonComponent = render(
       <Provider store={store}>
         <BrowserRouter>
-          <EventDetails />
+          <StatusButton
+            name={"Going"}
+            color={"#58D573"}
+            eventId={1}
+            eventStatus={"Maybe"}
+          />
         </BrowserRouter>
       </Provider>
     );
   });
 
-  test("EventDetails component renders", () => {
-    const firstDiv = document.querySelector(".event-details-container");
-    expect(firstDiv.toBeInDocument);
+  test("StatusButton component renders", () => {
+    expect(StatusButtonComponent.getByText(/Going/i));
   });
 });
