+++ src/components/events/view-events/ViewEvents.test.js
@@ -1,30 +1,29 @@
 import React from "react";
-import CalendarRow from "./CalendarRow.js";
+import ViewEvents from "./ViewEvents.js";
 import { render } from "@testing-library/react";
+import "@testing-library/jest-dom/extend-expect";
 import { BrowserRouter } from "react-router-dom";
 import { createStore, applyMiddleware } from "redux";
 import { Provider } from "react-redux";
-import { rootReducer } from "../../utilities/reducers";
+import { rootReducer } from "../../../utilities/reducers";
 import thunk from "redux-thunk";
-
-import "@testing-library/jest-dom/extend-expect";
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
 
-describe("Test CalendarRow static properties", () => {
-  let CalendarRowComponent;
+describe("Test view event static properties", () => {
+  let ViewEventsComponent;
   beforeEach(() => {
-    CalendarRowComponent = render(
+    ViewEventsComponent = render(
       <Provider store={store}>
         <BrowserRouter>
-          <CalendarRow />
+          <ViewEvents />
         </BrowserRouter>
       </Provider>
     );
   });
 
-  test("CalendarRow component renders", () => {
-    const firstDiv = document.querySelector(".calendar-row");
+  test("ViewEvents component renders", () => {
+    const firstDiv = document.querySelector(".middle-calendar");
     expect(firstDiv.toBeInDocument);
   });
 });
